+++ imager/settings.py
@@ -62,7 +62,7 @@ DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.postgresql_psycopg2',
         'NAME': 'djangoimager',
-        'USER': 'jakeanderson',
+        'USER': 'ndraper2',
     }
 }
